% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{as.mcmc}
\alias{as.mcmc}
\title{Coerce objects to coda mcmc objects}
\usage{
as.mcmc(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Further arguments passed to methods.}
}
\value{
An object, typically of class \code{"mcmc"}.
}
\description{
Generic function for converting objects to \code{"mcmc"} objects.
}
