% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-roc.R
\name{bayes_roc}
\alias{bayes_roc}
\title{Generic function for ROC curve extraction
Extracts ROC curve information from a fitted object.}
\usage{
bayes_roc(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
An object of class \code{"BayesODT_roc"}.
}
\description{
Generic function for ROC curve extraction
Extracts ROC curve information from a fitted object.
}
