% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesodt-package.R
\docType{package}
\name{bayesodt-package}
\alias{bayesodt}
\alias{bayesodt-package}
\title{bayesodt: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
