% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-plot.R
\name{plot.BayesODT_roc}
\alias{plot.BayesODT_roc}
\title{Plot method for BayesODT ROC objects}
\usage{
\method{plot}{BayesODT_roc}(
  x,
  add = FALSE,
  xlab = "False Positive Rate",
  ylab = "True Positive Rate",
  main = NULL,
  col = 1,
  lwd = 1,
  lty = 1,
  pch = 1
)
}
\arguments{
\item{x}{An object of class \code{"BayesODT_roc"}.}

\item{add}{Logical; if \code{TRUE}, add the ROC curve to an existing plot.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{main}{Plot title.}

\item{col}{Color for points/lines.}

\item{lwd}{Line width.}

\item{lty}{Line type.}

\item{pch}{Plotting symbol.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Plot ROC curves from an object returned by \code{bayes_roc()}.
}
