% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ABU.R
\name{plot_ABU}
\alias{plot_ABU}
\title{Plot of Posterior Samples}
\usage{
plot_ABU(fit)
}
\arguments{
\item{fit}{The fitted model object from BayesODT}
}
\value{
Posterior means and 95\% C.I.s of the parameters.
}
\description{
Plot posterior means and 95\% credible intervals for a_j, b_j, u_i parameters.
}
