% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ROC.R
\name{plot_ROC}
\alias{plot_ROC}
\title{ROC curves}
\usage{
plot_ROC(fit, method = c("Individual", "ROC1", "ROC2", "Smoothed.ROC",
  "Smoothed.pROC", "pROC"), ind = NULL)
}
\arguments{
\item{fit}{The fitted model object from BayesODT}

\item{method}{Select a method to produce an ROC curve ("Individual", "ROC1", "ROC2", "Smoothed.ROC", "Smoothed.pROC", "pROC")}

\item{ind}{For an individual ROC curve, indicate patient or subject ID}
}
\value{
Plot of ROC curves.
}
\description{
Produce ROC curves
}
