% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_trace}
\alias{plot_trace}
\title{Trace plot for BayesODT posterior samples
Plot an MCMC trace plot for selected posterior samples from a 'BayesODT' object using package 'coda'.}
\usage{
plot_trace(
  x,
  param = c("A", "a", "tau_a", "b", "mu_b", "tau_b", "u"),
  index = 1L,
  start = NULL,
  thin = NULL,
  legend_pos = "topright",
  ...
)
}
\arguments{
\item{x}{A fitted object of class \verb{BayesODT"}.}

\item{param}{Character string specifying which parameter to plot. One of \code{"A"}, \code{"a"}, \code{"tau_a"}, \code{"b"}, \code{"mu_b"}, \code{"tau_b"}, or \code{"u"}.}

\item{index}{Integer index (or vector of indices) of the selected component(s). For scalar hyperparameters such as \code{"tau_a"}, \code{"mu_b"}, and \code{"tau_b"}, only \code{index = 1} is valid.}

\item{start}{Optional starting iteration for the \code{coda::mcmc} object. If \code{NULL}, uses 1.}

\item{thin}{Optional thinning interval for the \code{coda::mcmc} object. If \code{NULL}, uses \code{x$mcmc_info$n.thin} when available, otherwise 1.}
}
\value{
Invisibly returns the created \code{"mcmc"} object or \code{"mcmc.list"} object.
}
\description{
Trace plot for BayesODT posterior samples
Plot an MCMC trace plot for selected posterior samples from a 'BayesODT' object using package 'coda'.
}
