% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-print.R
\name{print.BayesODT}
\alias{print.BayesODT}
\title{Print a BayesODT object}
\usage{
\method{print}{BayesODT}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"BayesODT"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Displays basic information about a fitted \code{"BayesODT"} object,
including data dimensions, missing-rating information, disease-status
setting, and MCMC information.
}
