% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-roc.R
\name{print.BayesODT_roc}
\alias{print.BayesODT_roc}
\title{Print a BayesODT ROC object}
\usage{
\method{print}{BayesODT_roc}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"BayesODT_roc"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Displays basic information about an object returned by \code{bayes_roc()}.
}
