% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{print.GoF.BayesODT}
\alias{print.GoF.BayesODT}
\title{Print Goodness-of-Fit Test Results}
\usage{
\method{print}{GoF.BayesODT}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{"GoF.BayesODT"} returned by \code{\link{GoF}}.}

\item{digits}{Number of digits to print.}

\item{...}{Additional arguments passed to \code{print}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for objects of class \code{"GoF.BayesODT"}.
}
