% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Simulated data for BayesODT}
\format{A list object with w and D:
\describe{
  \item{w}{90 (patients) by 50 (raters) classification matrix}
  \item{D}{true disease status for 90 patients}
}}
\source{
simulated dataset
}
\usage{
data(simdata)
}
\description{
A dataset containing simulated w (classification data) and D (true disease status)
}
\keyword{datasets}
