% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{Posterior Samples of the Parameters}
\usage{
summary(fit)
}
\arguments{
\item{fit}{The fitted model object from BayesODT}
}
\value{
aj Posterior samples of a_j parameters

bj Posterior samples of b_j parameters

ui Posterior samples of u_i parameters

alphak Posterior samples of alpha[k] parameters

aj.ci 95\% C.I. of a_j parameters

bj.ci 95\% C.I. of b_j parameters

ui.ci 95\% C.I. of c_i parameters

alphak.ci 95\% C.I. of alpha_k parameters
}
\description{
Obtain posterior samples of the parameters, a_j, b_j, u_i and alpha_k.
}
