% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-roc.R
\name{bayes_roc}
\alias{bayes_roc}
\title{Compute ROC curves from a fitted BayesODT object}
\usage{
bayes_roc(object, ...)
}
\arguments{
\item{object}{A fitted object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
An object representing ROC curve information.
}
\description{
Generic function for extracting ROC curve information from fitted objects.
}
