% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesODT.R
\name{BayesODT}
\alias{BayesODT}
\title{Fitting the Bayesain hierarchical model for ordinal classification processes}
\usage{
BayesODT(w = w, D = D, init = init, prior = prior, n.iter = 4000,
  n.burnin = 2000, n.thin = 5, ROC = FALSE)
}
\arguments{
\item{w}{The classification data (m: num. of patients by n: num. of raters) matrix}

\item{D}{The true disease status data}

\item{init}{Setting initial values for z, z0, a, b, u, A. If z0 is empty, the model will not use Model (1) in the manuscript.}

\item{prior}{Setting hyper-parameters (mu_a, tau_a, mu_b, tau_b, delta, gamma, tau, mu_u, tau_u)}

\item{n.iter}{Number of MCMC iterations}

\item{n.burnin}{Number of burnin iterations}

\item{n.thin}{Number of the thinning interval}

\item{ROC}{Optional. ROC=TRUE for plotting ROC curves. Default is FALSE.}
}
\value{
List of posterior samples of the parameters.
}
\description{
Fitting the Bayesain hierarchical model for ordinal classification processes
}
