% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoF.R
\name{GoF}
\alias{GoF}
\title{Goodness of Fit}
\usage{
GoF(fit)
}
\arguments{
\item{fit}{The fitted model object from BayesODT}
}
\value{
Estimated.Prob The estimated probabilities for each classification

Estimated.Prob.ci The 95\% C.I.s for the estimated probabilities

Empirical.Prob The empirical probabilities from the observed data
}
\description{
Check goodness of fit of the model.
}
