% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-roc.R
\name{print.BayesODT_roc}
\alias{print.BayesODT_roc}
\title{Print a BayesODT ROC object
Displays a concise summary of an object returned by \code{\link[=bayes_roc]{bayes_roc()}}.}
\usage{
\method{print}{BayesODT_roc}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"BayesODT_roc"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Print a BayesODT ROC object
Displays a concise summary of an object returned by \code{\link[=bayes_roc]{bayes_roc()}}.
}
